+++ lib/conjur/dsl2/planner/permissions.rb
@@ -1,5 +1,5 @@
 require 'conjur/dsl2/planner/base'
-require 'set'
+require 'conjur/dsl2/planner/facts'
 
 module Conjur
   module DSL2
@@ -11,64 +11,77 @@ module Conjur
       # privilege on an existing resource that is *not* given should be denied.
       class Permit < Base
         def do_plan
-          resources = Array(record.resources)
-          privileges = Array(record.privilege)
-          given_permissions = Hash.new { |hash, key| hash[key] = [] }
-          requested_permissions = Hash.new { |hash, key| hash[key] = [] }
-
-          resources.each do |resource|
-            permissions = begin
-              JSON.parse(api.resource(resource.resourceid).get)['permissions'] 
-            rescue RestClient::ResourceNotFound
-              []
-            end
-
-            permissions.each do |permission|
-              if privileges.member?(permission['privilege'])
-                given_permissions[[permission['privilege'], permission['resource']]].push [ permission['role'], permission['grant_option'] ]
-              end
-            end
-
-            privileges.each do |privilege|
-              Array(record.roles).each do |role|
-                requested_permissions[[privilege, resource.resourceid]].push [ role.role.roleid, !!role.admin ]
-              end
+          facts = PrivilegeFacts.new self
+          
+          facts.add_requested_permission record
+          
+          privileges = Array(record.privileges)
+          Array(record.resources).each do |resource|
+            facts.resource_permissions(resource, privileges) do |permission|
+              facts.add_existing_permission permission
             end
           end
-                      
-          resources.each do |resource|
-            error(%Q("Resource "#{resource}" not found in [#{plan.resources_created.to_a.sort.join(', ')}])) unless resource_exists?(resource)
+              
+          facts.validate!
 
-            privileges.each do |privilege|
-
-              target = resource.resourceid
-              given = given_permissions[[privilege, target]]
-              requested = requested_permissions[[privilege, target]]
-
-              (Set.new(requested) - Set.new(given)).each do |p|
-                role, admin = p
-
-                error(%Q(Role "#{role}" not found")) unless role_exists?(role)
-
-                permit = Conjur::DSL2::Types::Permit.new
-                permit.resource = resource_record target
-                permit.privilege = privilege
-                permit.role = Conjur::DSL2::Types::Member.new role_record(role)
-                permit.role.admin = true if admin
-                action permit
-              end
-
-              if record.replace
-                (Set.new(given) - Set.new(requested)).each do |p|
-                  role, admin = p
-                  deny = Conjur::DSL2::Types::Deny.new
-                  deny.resource = resource_record target
-                  deny.privilege = privilege
-                  deny.role = role_record(role)
-                  action deny
-                end
-              end
+          facts.grants_to_apply.each do |grant|
+            role, privilege, resource, admin = grant
+            
+            permit = Conjur::DSL2::Types::Permit.new
+            permit.resource = resource_record resource
+            permit.privilege = privilege
+            permit.role = Conjur::DSL2::Types::Member.new role_record(role)
+            permit.role.admin = true if admin
+            action permit
+          end          
+        
+          if record.replace
+            facts.grants_to_revoke.each do |grant|
+              roleid, privilege, resourceid = grant
+              deny = Conjur::DSL2::Types::Deny.new
+              deny.resource = resource_record resourceid
+              deny.privilege = privilege
+              deny.role = role_record(roleid)
+              action deny
             end
+          end
+        end
+      end
+      
+      # Plans a permission denial.
+      #
+      # A Deny statement is generated if the permission is currently held. Otherwise, its a nop.
+      class Deny < Base
+        def do_plan
+          facts = PrivilegeFacts.new self
+          
+          # Load all the permissions as both requested and existing grants.
+          # Then remove the Deny record, and see what's left.
+          privileges = Array(record.privileges)
+          Array(record.resources).each do |resource|
+            facts.resource_permissions(resource, privileges) do |permission|
+              permit_record = Types::Permit.new
+              permit_record.role = Types::Role.new(permission['role'])
+              permit_record.role.admin = permission['grant_option']
+              permit_record.privilege = permission['privilege']
+              permit_record.resource = Types::Resource.new(permission['resource'])
+              facts.add_requested_permission permit_record
+              
+              facts.add_existing_permission permission
+            end
+          end
+            
+          facts.remove_revoked_permission record
+          
+          facts.validate!
+          
+          facts.grants_to_revoke.each do |grant|
+            role, privilege, resource = grant
+            deny = Conjur::DSL2::Types::Deny.new
+            deny.resource = resource_record resource
+            deny.privilege = privilege
+            deny.role = role_record(role)
+            action deny
           end
         end
       end