+++ spec/spec_helper.rb
@@ -1,3 +1,9 @@
+require 'simplecov'
+SimpleCov.start do
+  add_filter '/spec/'
+  add_filter '/features/'
+end
+
 $LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
 require 'conjur-asset-dsl2'
 require 'logger'
@@ -7,6 +13,8 @@ if ENV['DEBUG']
   Conjur::DSL2::Planner::Base.logger.level = Logger::DEBUG
   Conjur::DSL2::Executor::Base.logger.level = Logger::DEBUG
 end
+
+Conjur::DSL2::Planner::BaseFacts.sort = true
 
 shared_context "planner", planning: true do
   let(:api) { double(:api) }