+++ CHANGELOG.md
@@ -1,6 +1,13 @@
+# 0.6.0
+
+* Implement the !deny statement.
+* Eliminate un-necessary privilege and role revocations.
+
 # 0.5.0
 
 * Refactor how the policy statements are validated and normalized, fixing some bugs in the process.
+* In record ids, replace the string '$namespace' with the policy namespace. This enables cross-policy
+  entitlements to be made more flexibly. 
 
 # 0.4.4